+++ lib/main.dart
@@ -3,11 +3,11 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutterapp/create_entry.dart';
-import 'package:flutterapp/initialPage.dart';
+import 'package:flutterapp/initial_page.dart';
 import 'package:intl/intl.dart';
 import 'diary.dart';
 import 'diary_detail.dart';
-import 'registerPage.dart';
+import 'register_page.dart';
 import 'package:rxdart/rxdart.dart';
 
 void main() => runApp(MyApp());
